barcodeclassifier/pyproject.toml
[project]
name = "barcodeclassifier"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "pillow>=11.2.1",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "tqdm>=4.67.1",
    "opencv-python>=4.11.0.86",
    "scikit-learn>=1.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/barcodeclassifier"]
